Numerous fixes
The list of genres in the sidebar now has a height cap, thus showing the
scrollbar. Added a lot of smaller fixes in terms of sending data
properly when closing modals for different reasons as well as added
handling of said modals when deleting something so that it also updates
the view accordingly.