Somewhat major change to the navbar
It now only uses one view and one controller, previously it used one for
offline users and one for online users, since we don't do that in other
areas anymore and I could solve it through ng-show, I deleted the
offline-version of it and edited the online version. This also solves a
couple of other problems we had with it completely switching view for
almost the same thing. (Radio buttons were acting up.)