Streamlined genre/author/book details.
All three now make use of a new object that is created using the latest
information from the point that you open up the modal. This also means
that if you edit a genre name you will not see the change in the sidebar
until you reload the page, this is for the moment intended behaviour as
all three now behaves the same, and it is now easier to predict their
behaviour as well as possibly fix it.