Keys of creation!
Dare you enter the temple of doom? You will need the three keys of
creation!

Aight enough joking around. ^^

The problem with the disappearing genres and authors was that the
success call that ran the loops that added them sometimes ran before the
authors/genres successcalls were completed, thus leaving us with a very
empty list.